+++ client/cody/CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to Sourcegraph Cody will be documented in this file.
 
 ### Fixed
 
+- Display error messages from compeletion client in webview instead of system messages
+
 ### Changed
 
 ## [0.0.10]